feat: remove API route fetching wrapper (#536)
* fix: cyclical entity store imports

* fix: import in test

* fix: use log instead of error for no env passed in

fix

* fix: don't use API fetch wrapper for now

Next seems to have problems in our middleware when using relative paths to call our API handlers. We need to pass an absolute path, but Next doesn't provide any mechanism of getting the absolute path on the server at request time.

* fix: update orphan relative call
