+++ apps/web/core/io/api/space.ts
@@ -1,6 +1,7 @@
+import { Environment } from '~/core/environment';
 import { Space } from '~/core/types';
 
-import { fetcher } from './geo-fetch';
+import { fetchSpace } from '../subgraph';
 
 interface NetworkData {
   space: Space | null;
@@ -8,5 +9,17 @@ interface NetworkData {
 }
 
 export async function space(spaceId: string): Promise<NetworkData> {
-  return fetcher<NetworkData>(`/api/space/${spaceId}`);
+  const config = Environment.getConfig(process.env.NEXT_PUBLIC_APP_ENV);
+  let space = await fetchSpace({ endpoint: config.subgraph, id: spaceId });
+  let isPermissionlessSpace = false;
+
+  if (!space) {
+    space = await fetchSpace({ endpoint: config.permissionlessSubgraph, id: spaceId });
+    if (space) isPermissionlessSpace = true;
+  }
+
+  return {
+    space,
+    isPermissionlessSpace,
+  };
 }
